A mark of trying delta time

+++ include/linux_parser.h
@@ -26,8 +26,8 @@ int TotalProcesses();
 int RunningProcesses();
 std::string OperatingSystem();
 std::string Kernel();
-long calc_activeTime(std::string userS, std::string niceS, std::string systemS, std::string irqS, std::string softirqS, std::string stealS);
-long calc_idleTime(std::string idleS, std::string iowaitS);
+long double calc_activeTime(std::string userS, std::string niceS, std::string systemS, std::string irqS, std::string softirqS, std::string stealS);
+long double calc_idleTime(std::string idleS, std::string iowaitS);
 
 
 // CPU
